Admit patient
Please make sure you run new version of db script!
Admit patient working and allocates to room E100.
Had to change the hospital system form we were using to be a
receptionist form. There is now a blank hospitalSystem form which will
be used by doctors and the med tech's.