Added validation to final reception text boxes
I'd like a way for it to get focus back on that field that caused the
error (even though multiple textboxes fire the same validation code).
But still not have it so bad that it's an infinite loop until you
complete it correctly.

Maybe it needs to have or be 'on submit' final validation check.